user-failure-analysis/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
plotly==5.18.0
python-dotenv==1.0.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-functions==1.17.0
python-multipart==0.0.6
pydantic==2.5.1 